add automated data ingest in various formats
Also ingest, converter, collator factories, generic readers, etc.


Reader: Takes a filepath and reads it into a pandas dataframe with (hopefully) the right required fields. Works for various formats of data including astropy tables. Identifies time columns.

Converter: dataclass that stores header and other information for a file source to convert its read dataframe it into a Photometry object.

Collator: Takes a path and a reader and a Converter and creates Photometry object from all files defined by the Converter standard and can be read by the reader. Can also be defined for upper limit data.

We have several predefined Converters and Collators, for flows and ztf, as well as generic ECSV,CSV,JSON ones.

DataIngestor: Takes a path, set of Collators (or one) by default using the built-in ones, and creates a SN object with all the photometry and the predefined data. It can also uses public APIs to fill in missing fields. It really only needs the SN name and can use `tnsquery` to fill in the rest.